# V2: improve performance, remove dependencies, code refactoring