--- a/.travis.yml
+++ b/.travis.yml
@@ -3,9 +3,6 @@ node_js:
   - "stable"
   - "10"
   - "8"
-  - "6"
-  - "5"
-  - "4"
 script:
   - npm run lint
   - npm test