--- a/benchmark/index.js
+++ b/benchmark/index.js
@@ -23,32 +23,32 @@ const data = [
       lastName: 'Maggio',
       email: 'Kaley-Maggio@example.com',
       age: '80',
-      gender: 'male',
+      gender: 'male'
     },
     tags: ['1', '2'],
     photos: [
       'ed70cf44-9a34-4878-84e6-0c0e4a450cfe',
       '24ba3666-a593-498c-9f5d-55a4ee08c72e',
-      'f386492d-df61-4573-b4e3-54f6f5d08acf',
+      'f386492d-df61-4573-b4e3-54f6f5d08acf'
     ],
     comments: [
       {
         _id: '1',
         body: 'First !',
-        created: '2015-08-14T18:42:16.475Z',
+        created: '2015-08-14T18:42:16.475Z'
       },
       {
         _id: '2',
         body: 'I Like !',
-        created: '2015-09-14T18:42:12.475Z',
+        created: '2015-09-14T18:42:12.475Z'
       },
       {
         _id: '3',
         body: 'Awesome',
-        created: '2015-09-15T18:42:12.475Z',
-      },
-    ],
-  },
+        created: '2015-09-15T18:42:12.475Z'
+      }
+    ]
+  }
 ];
 
 serializer.register('article', {
@@ -58,7 +58,7 @@ serializer.register('article', {
     self(d) {
       // Can be a function or a string value ex: { self: '/articles/1'}
       return `/articles/${d.id}`;
-    },
+    }
   },
   relationships: {
     // An object defining some relationships.
@@ -68,32 +68,32 @@ serializer.register('article', {
         // An object or a function that describes Relationships links
         return {
           self: `/articles/${d.id}/relationships/author`,
-          related: `/articles/${d.id}/author`,
+          related: `/articles/${d.id}/author`
         };
-      },
+      }
     },
     tags: {
-      type: 'tag',
+      type: 'tag'
     },
     photos: {
-      type: 'photo',
+      type: 'photo'
     },
     comments: {
       type: 'comment',
-      schema: 'only-body', // A custom schema
-    },
+      schema: 'only-body' // A custom schema
+    }
   },
   topLevelMeta(d, extraData) {
     // An object or a function that describes top level meta.
     return {
       count: extraData.count,
-      total: d.length,
+      total: d.length
     };
   },
   topLevelLinks: {
     // An object or a function that describes top level links.
-    self: '/articles', // Can be a function (with extra data argument) or a string value
-  },
+    self: '/articles' // Can be a function (with extra data argument) or a string value
+  }
 });
 
 // Register 'people' type
@@ -102,23 +102,23 @@ serializer.register('people', {
   links: {
     self(d) {
       return `/peoples/${d.id}`;
-    },
-  },
+    }
+  }
 });
 
 // Register 'tag' type
 serializer.register('tag', {
-  id: 'id',
+  id: 'id'
 });
 
 // Register 'photo' type
 serializer.register('photo', {
-  id: 'id',
+  id: 'id'
 });
 
 // Register 'comment' type with a custom schema
 serializer.register('comment', 'only-body', {
-  id: '_id',
+  id: '_id'
 });
 
 let serialized;
@@ -131,13 +131,18 @@ console.log(`${os.type()} ${os.release()} ${os.arch()}`);
 console.log('Node.JS:', process.versions.node);
 console.log('V8:', process.versions.v8);
 
-let cpus = os.cpus().map(cpu => cpu.model).reduce((o, model) => {
-  if (!o[model]) o[model] = 0;
-  o[model] += 1;
-  return o;
-}, {});
+let cpus = os
+  .cpus()
+  .map(cpu => cpu.model)
+  .reduce((o, model) => {
+    if (!o[model]) o[model] = 0;
+    o[model] += 1;
+    return o;
+  }, {});
 
-cpus = Object.keys(cpus).map(key => `${key} \u00d7 ${cpus[key]}`).join('\n');
+cpus = Object.keys(cpus)
+  .map(key => `${key} \u00d7 ${cpus[key]}`)
+  .join('\n');
 
 console.info(cpus);
 
@@ -147,8 +152,10 @@ suite
   .add('serializeAsync', {
     defer: true,
     fn(deferred) {
-      serializer.serializeAsync('article', data, { count: 2 }).then(() => { deferred.resolve(); });
-    },
+      serializer.serializeAsync('article', data, { count: 2 }).then(() => {
+        deferred.resolve();
+      });
+    }
   })
   .add('serialize', () => {
     serialized = serializer.serialize('article', data, { count: 2 });
@@ -156,8 +163,10 @@ suite
   .add('deserializeAsync', {
     defer: true,
     fn(deferred) {
-      serializer.deserializeAsync('article', serialized).then(() => { deferred.resolve(); });
-    },
+      serializer.deserializeAsync('article', serialized).then(() => {
+        deferred.resolve();
+      });
+    }
   })
   .add('deserialize', () => {
     serializer.deserialize('article', serialized);
@@ -174,16 +183,15 @@ suite
       status: '422',
       source: { pointer: '/data/attributes/error' },
       title: 'Error',
-      detail: 'An error occured',
+      detail: 'An error occured'
     };
 
     serializer.serializeError(jsonapiError);
   })
   // add listeners
-  .on('cycle', (event) => {
+  .on('cycle', event => {
     console.log(String(event.target));
   })
-  .on('complete', () => {
-  })
-// run  async
+  .on('complete', () => {})
+  // run  async
   .run({ async: false });