--- a/package.json
+++ b/package.json
@@ -27,18 +27,14 @@
     "bson-objectid": "^1.1.4",
     "chai": "^4.1.2",
     "coveralls": "^3.0.1",
-    "eslint": "^4.9.0",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-plugin-import": "^2.7.0",
+    "eslint": "^5.16.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.17.2",
     "istanbul": "^0.4.2",
     "mocha": "^5.2.0"
   },
   "dependencies": {
-    "into-stream": "^3.1.0",
-    "joi": "^12.0.0",
-    "lodash": "^4.5.1",
-    "stream-to-array": "^2.3.0",
-    "through2": "^2.0.3",
-    "unique-stream": "^2.2.1"
+    "30-seconds-of-code": "^1.2.3",
+    "lodash.set": "^4.3.2"
   }
 }