--- a/package.json
+++ b/package.json
@@ -29,9 +29,12 @@
     "coveralls": "^3.0.1",
     "eslint": "^5.16.0",
     "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-config-prettier": "^4.2.0",
     "eslint-plugin-import": "^2.17.2",
+    "eslint-plugin-prettier": "^3.0.1",
     "istanbul": "^0.4.2",
-    "mocha": "^5.2.0"
+    "mocha": "^5.2.0",
+    "prettier": "^1.17.0"
   },
   "dependencies": {
     "30-seconds-of-code": "^1.2.3",