--- a/test/helpers/tick-counter.js
+++ b/test/helpers/tick-counter.js
@@ -1,5 +1,3 @@
-'use strict';
-
 class TickCounter {
   constructor(max) {
     this.ticks = 0;