… lib/JSONAPISerializer.js
… test/unit/validator.test.js