### Goals
- improve performance (add benchmark) #63 
- remove dependencies (lodash, joi) #63 
- code refactoring
- Keep same API as v1

### Breaking change
The only breaking change will be incompatibility with older version of Nodejs < 8

### Benchmark comparison
```bash
Platform info:
==============
Darwin 18.5.0 x64
Node.JS: 10.15.2
V8: 6.8.275.32-node.12
Intel(R) Core(TM) i7-4770HQ CPU @ 2.20GHz × 8

Suite V1:
==============
serializeAsync x 5,154 ops/sec ±9.10% (72 runs sampled)
serialize x 28,150 ops/sec ±1.22% (86 runs sampled)
deserializeAsync x 49,116 ops/sec ±1.80% (81 runs sampled)
deserialize x 148,658 ops/sec ±1.26% (92 runs sampled)
serializeError x 281,786 ops/sec ±0.99% (90 runs sampled)
serializeError with a JSON API error object x 10,637 ops/sec ±2.82% (85 runs sampled)

Suite V2:
==============
serializeAsync x 64,411 ops/sec ±0.59% (82 runs sampled)
serialize x 150,408 ops/sec ±0.63% (93 runs sampled)
deserializeAsync x 111,881 ops/sec ±0.31% (84 runs sampled)
deserialize x 445,711 ops/sec ±0.29% (94 runs sampled)
serializeError x 294,273 ops/sec ±0.79% (90 runs sampled)
serializeError with a JSON API error object x 16,864,514 ops/sec ±0.32% (93 runs sampled)
```