@danivek 2019-04-30
@mattiloh, I try an other implementation of non blocking (de)serialization (removing usage of stream) to improve performance. You can see it here: https://github.com/danivek/json-api-serializer/blob/v2/lib/JSONAPISerializer.js#L325. All tests are passing.

Do you see any issue with this implementation ?