@mattiloh 2019-04-30
Wow! Really great performance improvements 🚀 
Is it coming from the usage of `Map` instead of an array for `included`? 🤔

Sure, I'm fine with using `setImmediate` instead of streams. Unfortunately, it's still not working in browsers with it, since `setImmediate` seems [not to be really widely supported](https://caniuse.com/#feat=setimmediate) yet. But I guess it's a step in the right direction.