@danivek 2019-04-30
Yes a little bit, but the real performance gain comes from the use of native functions rather than lodash ones + some code optimization.

Of course this library still not work on browsers because of `setImmediate`. Maybe we need to fallback to `setTimeout` or use polyfill to make it work.