@johannesschobel 2019-04-30
Hey @danivek , good to see the improvements.. nice work on the performance!
May I add another feature request for the upcoming version? I would really (!!!) love to see "official typescript" definitions. Maybe you can add those as well?

Thank you very much in advance!