@mattiloh 2019-04-30
> Of course this library still not work on browsers because of setImmediate. Maybe we need to fallback to setTimeout or use polyfill to make it work.

So far, I didn't need it in the browser, so no hard feelings from my side 😉 
But I think it has been requested sometimes. If `setImmediate` is the only problem for browser-support, we could also think about adding a section to the README to explain how it could be used with a `setImmediate` polyfill.
Anyhow, I would suggest to tackle browser support in a separate PR (if it's really needed).