@johannesschobel 2019-04-30
Hey @danivek ,

i would like to propse another feature request.. 
What do you think about implementing something like "Transformers" known from `Fractal` (php package for transforming & serializing data into various formats).
https://fractal.thephpleague.com/

Using Transformers would allow for:
1) easy rename properties: because they take the part of "mapping" from one structure to another one
2) allow for computed properties: because you would be able to add new properties to this transformer
3) convert types: e.g., you would be able to convert values to a proper type (e.g., the `id` **must** be a string)
4) ...

What do you think about this?
All the best