@danivek 2019-04-30
I think we should avoid introducing a new kind of domain-specific language (DSL) in this library. `json-api-serializer` is only responsible to (de)serialize a data to JSON API compliant responses. If you want to transform your data, you should do it before using this library.

The only possible option we could add for this library, is a `transform` option, that will act as a hook before (de)serializing attributes. So you will be able to use any kind of library you want to transform attributes.