@johannesschobel 2019-04-30
thanks for your response @danivek ... i think, a hook to transform the data would be nice.. Any idea how this would / could look like? Would it be be possible to just call a method and pass the current object?

Like this:
```
// user.transformer.ts
class UserTransformer extends Transformer() {
  transform(entity: User): UserResponse {
    return {
      id: String(entity.id),
      firstName: entity.firstname,
      // ...
      name: entity.firstname + ' ' + entity.lastname,
    }
  }
}
```

What do you think?